Digital-wardrobe-main/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-multipart
mangum
pydantic
